feat(users): apply vertical slice architecture to users
