--- src/core/users/domain/interfaces/users-repository.ts
@@ -1,5 +1,0 @@
-import type { User } from "@/src/core/users/domain/models/user";
-
-export interface IUsersRepository {
-  users(): Promise<Array<User>>;
-}
