+++ src/ui/@types/next-intl.d.ts
@@ -1,5 +1,6 @@
 interface Messages {
   home: typeof import("../i18n/messages/en/home.json");
   posts: typeof import("../i18n/messages/en/posts.json");
+  users: typeof import("../i18n/messages/en/users.json");
 }
 declare interface IntlMessages extends Messages {}
