+++ src/ui/features/misc/containers/login-button/login-button.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { Button } from "@/src/ui/components/button/button";
-import { useUserProvider } from "@/src/ui/providers/user.provider";
+import { useUserStore } from "@/src/users/application/stores/user-store";
 import React, { useEffect } from "react";
 import { LoggingModal } from "@/src/ui/components/logging-modal/logging-modal";
 import { CookieUtils } from "@front_web_mrmilu/utils";
@@ -8,10 +8,10 @@ import { useUiProvider } from "@/src/ui/providers/ui.provider";
 import { useSearchParams } from "next/navigation";
 
 export default function LoginButton() {
-  const setLogged = useUserProvider((state) => state.setLogged);
-  const login = useUserProvider((state) => state.login);
-  const logout = useUserProvider((state) => state.logout);
-  const isUserLogged = useUserProvider((state) => state.logged);
+  const setLogged = useUserStore((state) => state.setLogged);
+  const login = useUserStore((state) => state.login);
+  const logout = useUserStore((state) => state.logout);
+  const isUserLogged = useUserStore((state) => state.logged);
   const showModal = useUiProvider((state) => state.showModal);
   const params = useSearchParams();
   const protectedRouteAccessAttempt = params?.get("protectedRouteAccessAttempt");
