--- src/ui/features/users/state/users-list.provider.ts
@@ -1,6 +1,0 @@
-import { createProvider } from "@/src/common/utils/zustand";
-import type { UsersListStateViewModel } from "@/src/ui/features/users/view-models/users-list-state";
-
-export const useUsersListProvider = createProvider<UsersListStateViewModel>(() => () => ({
-  users: []
-}));
