--- src/ui/features/users/view-models/users-list-state.ts
@@ -1,5 +1,0 @@
-import type { User } from "@/src/core/users/domain/models/user";
-
-export interface UsersListStateViewModel {
-  users: Array<User>;
-}
