+++ src/ui/i18n/messages/en/users.json
@@ -1,0 +1,4 @@
+{
+    "title": "Users SSR Page",
+    "withProviderTitle": "Users with provider SSR page"
+}
