+++ src/ui/i18n/messages/es/users.json
@@ -1,0 +1,4 @@
+{
+    "title": "Página de Usuarios SSR",
+    "withProviderTitle": "Página de Usuarios con provider SSR"
+}
