+++ src/users/domain/interfaces/users-repository.ts
@@ -1,0 +1,5 @@
+import type { User } from "@/src/users/domain/models/user";
+
+export interface IUsersRepository {
+  users(): Promise<Array<User>>;
+}
