+++ src/users/presentation/view-models/user-state.ts
@@ -1,0 +1,6 @@
+export interface UserStateViewModel {
+  logged: boolean;
+  setLogged(logged: boolean): void;
+  logout(): void;
+  login(): void;
+}
