+++ src/users/presentation/view-models/users-list-state.ts
@@ -1,0 +1,5 @@
+import type { User } from "@/src/users/domain/models/user";
+
+export interface UsersListStateViewModel {
+  users: Array<User>;
+}
